feat(zac): avoid ZAC deleting compdump on reload
zsh-autocomplete includes functionality which tries to detect when it
gets reinitialized after another plugin like autocd has loaded in as
completion for -command-. If this is the case, ZAC will delete the
current zcompdump file to force a full reload. This is generally not a
problem since our reload procedure involves always deleting the compdump
file, however in the future we will want more finegrained control on
when full reloads happen. This commit ensures this.